requirements.txt
absl-py==0.11.0
anyio==2.0.2
argon2-cffi==20.1.0
astunparse==1.6.3
async-generator==1.10
attrs==19.3.0
Babel==2.9.0
backcall==0.2.0
bleach==3.2.2
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.4
cftime==1.4.1
chardet==4.0.0
chex==0.0.4
click==7.1.2
click-aliases==1.0.1
click-completion==0.5.2
coverage==5.4
cycler==0.10.0
dataclasses==0.6
decorator==4.4.2
defusedxml==0.6.0
dm-haiku==0.0.3
dm-tree==0.1.5
entrypoints==0.3
fhi-vibes==1.0.3
flatbuffers==1.12
future==0.18.2
gast==0.4.0
google-auth==1.28.0
google-auth-oauthlib==0.4.3
google-pasta==0.2.0
grpcio==1.34.1
gviz-api==1.9.0
h5py==3.1.0
idna==2.10
iniconfig==1.1.1
ipykernel==5.4.3
ipython==7.19.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
jax==0.2.7
jax-md==0.1.15
jaxlib==0.1.57
jconfigparser==0.1.3
jedi==0.18.0
Jinja2==2.11.2
json5==0.9.5
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.11
jupyter-console==6.2.0
jupyter-core==4.7.0
jupyter-server==1.2.2
jupyterlab==3.0.5
jupyterlab-pygments==0.1.2
jupyterlab-server==2.1.2
jupyterlab-widgets==1.0.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.4
MarkupSafe==1.1.1
matplotlib==3.3.3
mistune==0.8.4
nbclassic==0.2.6
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.4.3
netCDF4==1.5.6
notebook==6.2.0
numexpr==2.7.3
numpy==1.20.1
oauthlib==3.1.0
opt-einsum==3.3.0
packaging==20.8
pandas==1.2.3
pandocfilters==1.4.3
parso==0.8.1
pexpect==4.8.0
phonopy==2.8.1
pickleshare==0.7.5
Pillow==8.1.0
pluggy==0.13.1
prometheus-client==0.9.0
prompt-toolkit==3.0.14
protobuf==3.15.6
ptyprocess==0.7.0
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.7.4
pyparsing==2.4.7
pyrsistent==0.17.3
pytest==6.2.2
python-dateutil==2.8.1
pytz==2020.5
PyYAML==5.4.1
pyzmq==21.0.2
qtconsole==5.0.2
QtPy==1.9.0
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scipy==1.5.4
seaborn==0.11.1
seekpath==1.9.7
Send2Trash==1.5.0
shellingham==1.4.0
six==1.15.0
sniffio==1.2.0
son==0.3.3
spglib==1.16.1
tables==3.6.1
tensorboard-data-server==0.4.0
termcolor==1.1.0
terminado==0.9.2
testpath==0.4.4
tf-estimator-nightly==2.5.0.dev2021031901
toml==0.10.2
toolz==0.11.1
tornado==6.1
traitlets==5.0.5
typing-extensions==3.7.4.3
urllib3==1.26.2
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
xarray==0.17.0
